move AsyncBlockIterator to block_iterator.rs
