+++ storage/src/lsm_tree/iterator/block_iterator.rs
@@ -200,12 +200,49 @@ impl BlockIterator {
 }
 
 #[cfg(test)]
-mod tests {
+pub mod tests {
+    use async_trait::async_trait;
     use test_log::test;
 
     use super::*;
     use crate::components::{BlockBuilder, BlockBuilderOptions};
+    use crate::iterator::Iterator;
     use crate::utils::full_key;
+
+    pub struct AsyncBlockIterator(BlockIterator);
+
+    impl AsyncBlockIterator {
+        pub fn new(block: Arc<Block>) -> Self {
+            AsyncBlockIterator(BlockIterator::new(block))
+        }
+    }
+
+    #[async_trait]
+    impl Iterator for AsyncBlockIterator {
+        async fn next(&mut self) -> Result<()> {
+            self.0.next()
+        }
+
+        async fn prev(&mut self) -> Result<()> {
+            self.0.prev()
+        }
+
+        fn key(&self) -> &[u8] {
+            self.0.key()
+        }
+
+        fn value(&self) -> &[u8] {
+            self.0.value()
+        }
+
+        fn is_valid(&self) -> bool {
+            self.0.is_valid()
+        }
+
+        async fn seek<'s>(&mut self, seek: Seek<'s>) -> Result<bool> {
+            self.0.seek(seek)
+        }
+    }
 
     fn build_iterator_for_test() -> BlockIterator {
         let options = BlockBuilderOptions::default();
